+++ KeePKCS11/Forms/FormEnterLabelAndPIN.Designer.cs
@@ -42,10 +42,11 @@
             // checkBoxDisplayPIN
             // 
             this.checkBoxDisplayPIN.AutoSize = true;
-            this.checkBoxDisplayPIN.Location = new System.Drawing.Point(18, 133);
+            this.checkBoxDisplayPIN.Location = new System.Drawing.Point(24, 164);
+            this.checkBoxDisplayPIN.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.checkBoxDisplayPIN.Name = "checkBoxDisplayPIN";
-            this.checkBoxDisplayPIN.Size = new System.Drawing.Size(81, 17);
-            this.checkBoxDisplayPIN.TabIndex = 20;
+            this.checkBoxDisplayPIN.Size = new System.Drawing.Size(100, 20);
+            this.checkBoxDisplayPIN.TabIndex = 2;
             this.checkBoxDisplayPIN.Text = "Display PIN";
             this.checkBoxDisplayPIN.UseVisualStyleBackColor = true;
             this.checkBoxDisplayPIN.CheckedChanged += new System.EventHandler(this.checkBoxDisplayPIN_CheckedChanged);
@@ -53,28 +54,31 @@
             // label1
             // 
             this.label1.AutoSize = true;
-            this.label1.Location = new System.Drawing.Point(15, 90);
+            this.label1.Location = new System.Drawing.Point(20, 111);
+            this.label1.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
             this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(79, 13);
+            this.label1.Size = new System.Drawing.Size(95, 16);
             this.label1.TabIndex = 19;
             this.label1.Text = "Enter your PIN:";
             // 
             // tbxPinCode
             // 
-            this.tbxPinCode.Location = new System.Drawing.Point(18, 106);
+            this.tbxPinCode.Location = new System.Drawing.Point(24, 130);
+            this.tbxPinCode.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.tbxPinCode.Name = "tbxPinCode";
-            this.tbxPinCode.Size = new System.Drawing.Size(269, 20);
-            this.tbxPinCode.TabIndex = 16;
+            this.tbxPinCode.Size = new System.Drawing.Size(357, 22);
+            this.tbxPinCode.TabIndex = 1;
             this.tbxPinCode.UseSystemPasswordChar = true;
             this.tbxPinCode.KeyDown += new System.Windows.Forms.KeyEventHandler(this.tbxPinCode_KeyDown);
             // 
             // btnOk
             // 
             this.btnOk.DialogResult = System.Windows.Forms.DialogResult.OK;
-            this.btnOk.Location = new System.Drawing.Point(128, 160);
+            this.btnOk.Location = new System.Drawing.Point(171, 197);
+            this.btnOk.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.btnOk.Name = "btnOk";
-            this.btnOk.Size = new System.Drawing.Size(75, 23);
-            this.btnOk.TabIndex = 18;
+            this.btnOk.Size = new System.Drawing.Size(100, 28);
+            this.btnOk.TabIndex = 3;
             this.btnOk.Text = "&OK";
             this.btnOk.UseVisualStyleBackColor = true;
             this.btnOk.Click += new System.EventHandler(this.btnOk_Click);
@@ -82,10 +86,11 @@
             // btnCancel
             // 
             this.btnCancel.DialogResult = System.Windows.Forms.DialogResult.Cancel;
-            this.btnCancel.Location = new System.Drawing.Point(209, 160);
+            this.btnCancel.Location = new System.Drawing.Point(279, 197);
+            this.btnCancel.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.btnCancel.Name = "btnCancel";
-            this.btnCancel.Size = new System.Drawing.Size(75, 23);
-            this.btnCancel.TabIndex = 17;
+            this.btnCancel.Size = new System.Drawing.Size(100, 28);
+            this.btnCancel.TabIndex = 4;
             this.btnCancel.Text = "&Cancel";
             this.btnCancel.UseVisualStyleBackColor = true;
             this.btnCancel.Click += new System.EventHandler(this.btnCancel_Click);
@@ -93,33 +98,36 @@
             // label2
             // 
             this.label2.AutoSize = true;
-            this.label2.Location = new System.Drawing.Point(15, 21);
+            this.label2.Location = new System.Drawing.Point(20, 26);
+            this.label2.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
             this.label2.Name = "label2";
-            this.label2.Size = new System.Drawing.Size(92, 13);
+            this.label2.Size = new System.Drawing.Size(114, 16);
             this.label2.TabIndex = 22;
             this.label2.Text = "Enter object label:";
             // 
             // tbxObjectLabel
             // 
-            this.tbxObjectLabel.Location = new System.Drawing.Point(18, 57);
+            this.tbxObjectLabel.Location = new System.Drawing.Point(24, 70);
+            this.tbxObjectLabel.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.tbxObjectLabel.Name = "tbxObjectLabel";
-            this.tbxObjectLabel.Size = new System.Drawing.Size(269, 20);
-            this.tbxObjectLabel.TabIndex = 21;
+            this.tbxObjectLabel.Size = new System.Drawing.Size(357, 22);
+            this.tbxObjectLabel.TabIndex = 0;
             // 
             // label3
             // 
             this.label3.AutoSize = true;
-            this.label3.Location = new System.Drawing.Point(15, 38);
+            this.label3.Location = new System.Drawing.Point(20, 47);
+            this.label3.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
             this.label3.Name = "label3";
-            this.label3.Size = new System.Drawing.Size(234, 13);
+            this.label3.Size = new System.Drawing.Size(292, 16);
             this.label3.TabIndex = 23;
             this.label3.Text = "( leave this field empty to create it automatically )";
             // 
             // FormEnterLabelAndPIN
             // 
-            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(302, 205);
+            this.ClientSize = new System.Drawing.Size(403, 252);
             this.Controls.Add(this.label3);
             this.Controls.Add(this.label2);
             this.Controls.Add(this.tbxObjectLabel);
@@ -130,6 +138,7 @@
             this.Controls.Add(this.btnCancel);
             this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedSingle;
             this.Icon = ((System.Drawing.Icon)(resources.GetObject("$this.Icon")));
+            this.Margin = new System.Windows.Forms.Padding(4, 4, 4, 4);
             this.MaximizeBox = false;
             this.MinimizeBox = false;
             this.Name = "FormEnterLabelAndPIN";